re-factor code to avoid error when build the pkg
